Refactored all code into different folders.
Staff Profile screens are in their own folder.
Patient Profile screens are in their own folder.
Components used throughout both screens are now in their own folder.
